refactor: move pages into lib
