+++ packages/app/src/pages/_document.tsx
@@ -17,28 +17,7 @@
  * along with Essencium Frontend. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { createGetInitialProps } from '@mantine/next'
-import Document, { Head, Html, Main, NextScript } from 'next/document'
-
-const getInitialProps = createGetInitialProps()
+import { Document } from '@frachtwerk/essencium-lib'
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export default class _Document extends Document {
-  static getInitialProps = getInitialProps
-
-  render(): JSX.Element {
-    return (
-      <Html lang="en">
-        <Head>
-          <link rel="icon" href="/img/web/favicon.ico" sizes="any" />
-          <link rel="apple-touch-icon" href="/img/web/apple-touch-icon.png" />
-        </Head>
-        <body>
-          <div id="notification" />
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    )
-  }
-}
+export default Document
