+++ packages/app/src/pages/login/index.tsx
@@ -17,83 +17,9 @@
  * along with Essencium Frontend. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-  getTranslation,
-  Login,
-  LoginForm,
-  PublicLayout,
-  useCreateToken,
-  useResetPassword,
-  withBaseStylingShowNotification,
-} from '@frachtwerk/essencium-lib'
-import { ResetPassword } from '@frachtwerk/essencium-types'
-import { useRouter } from 'next/router'
-import { useTranslation } from 'next-i18next'
-import { useState } from 'react'
+import { LoginView } from '@frachtwerk/essencium-lib'
 
 import { baseGetStaticProps } from '@/utils/next'
-
-function LoginView(): JSX.Element {
-  const { t } = useTranslation()
-
-  const router = useRouter()
-
-  const { mutate: resetPassword, isLoading: isResettingPassword } =
-    useResetPassword()
-  const [isResetPasswordSent, setIsResetPasswordSent] = useState(false)
-  const [isPasswordResetFormOpened, setIsPasswordResetFormOpened] =
-    useState(false)
-
-  const { mutate: createToken } = useCreateToken()
-
-  function handleLogin(username: string, password: string): void {
-    createToken(
-      { username, password },
-      {
-        onSuccess: () => router.push('/'),
-        onError: () => {
-          withBaseStylingShowNotification({
-            title: t('loginView.errorMessage.title'),
-            color: 'error',
-            notificationType: 'created',
-          })
-        },
-      },
-    )
-  }
-
-  function handlePasswordReset(email: ResetPassword['email']): void {
-    resetPassword(email, {
-      onSuccess: () => {
-        setIsResetPasswordSent(true)
-        setIsPasswordResetFormOpened(false)
-      },
-    })
-  }
-
-  return (
-    <Login
-      form={
-        <LoginForm
-          handleLogin={handleLogin}
-          handlePasswordReset={handlePasswordReset}
-          setIsPasswordResetFormOpened={setIsPasswordResetFormOpened}
-          isResetPasswordSent={isResetPasswordSent}
-          isPasswordResetFormOpened={isPasswordResetFormOpened}
-          isResettingPassword={isResettingPassword}
-        />
-      }
-    />
-  )
-}
-
-LoginView.getLayout = function getLayout(page: React.ReactNode): JSX.Element {
-  return (
-    <PublicLayout routeName={getTranslation('loginView.title')}>
-      {page}
-    </PublicLayout>
-  )
-}
 
 export const getStaticProps = baseGetStaticProps()
