+++ packages/app/src/pages/profile/index.tsx
@@ -17,100 +17,9 @@
  * along with Essencium Frontend. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-  AuthLayout,
-  getTranslation,
-  logout,
-  Profile,
-  useGetMe,
-  useGetRoles,
-  useUpdateMe,
-  useUpdatePassword,
-} from '@frachtwerk/essencium-lib'
-import { PasswordChange, UserUpdate } from '@frachtwerk/essencium-types'
-import { Center, Loader } from '@mantine/core'
-import { useRouter } from 'next/router'
-import { useTranslation } from 'next-i18next'
-import { ReactElement } from 'react'
+import { ProfileView } from '@frachtwerk/essencium-lib'
 
 import { baseGetStaticProps } from '@/utils/next'
-
-function ProfileView(): JSX.Element {
-  const { t } = useTranslation()
-
-  const router = useRouter()
-
-  const { data: user, isLoading: isLoadingUser } = useGetMe()
-
-  const { mutate: updateUser } = useUpdateMe()
-
-  const { mutate: updatePassword } = useUpdatePassword()
-
-  const { data: rolesRequest } = useGetRoles({ page: 0, size: 9999 })
-
-  const roles = rolesRequest?.content || []
-
-  function handleUpdate(updatedUser: UserUpdate): void {
-    updateUser(updatedUser)
-
-    router.push('/profile', undefined, { locale: updatedUser.locale })
-  }
-
-  function handlePasswordUpdate(
-    password: PasswordChange['password'],
-    verification: PasswordChange['password'],
-  ): void {
-    updatePassword(
-      { password, verification },
-      {
-        onSuccess: () => {
-          logout()
-
-          router.push('/login')
-        },
-      },
-    )
-  }
-
-  if (user) {
-    return (
-      <Profile
-        user={user}
-        roles={roles}
-        handleUpdate={handleUpdate}
-        handlePasswordUpdate={handlePasswordUpdate}
-      />
-    )
-  }
-
-  if (isLoadingUser) {
-    return (
-      <Center h="100%">
-        <Loader size="xl" name="loader" />
-      </Center>
-    )
-  }
-
-  return (
-    <Center w="100%" h="100%">
-      {t('profileView.userNotFound')}
-    </Center>
-  )
-}
-
-ProfileView.getLayout = function getLayout(
-  page: ReactElement,
-  version?: string,
-) {
-  return (
-    <AuthLayout
-      routeName={getTranslation('profileView.title')}
-      version={version}
-    >
-      {page}
-    </AuthLayout>
-  )
-}
 
 export const getStaticProps = baseGetStaticProps()
