+++ packages/app/src/pages/users/add/index.tsx
@@ -17,89 +17,9 @@
  * along with Essencium Frontend. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-  AuthLayout,
-  getTranslation,
-  useCreateUser,
-  useGetRoles,
-  UserForm,
-  useZodForm,
-} from '@frachtwerk/essencium-lib'
-import { UserInput, userInputSchema } from '@frachtwerk/essencium-types'
-import { Card, Flex, Text, Title } from '@mantine/core'
-import { IconUserPlus } from '@tabler/icons-react'
-import { useRouter } from 'next/router'
-import { useTranslation } from 'next-i18next'
+import { AddUserView } from '@frachtwerk/essencium-lib'
 
-import { FORM_DEFAULTS } from '@/pages/users'
 import { baseGetStaticProps } from '@/utils/next'
-
-function AddUserView(): JSX.Element {
-  const router = useRouter()
-
-  const { t } = useTranslation()
-
-  const { handleSubmit, control, formState, setValue } = useZodForm({
-    schema: userInputSchema,
-    defaultValues: FORM_DEFAULTS,
-  })
-
-  const { mutate: addUser } = useCreateUser()
-
-  const { data: rolesResponse } = useGetRoles({
-    page: 0,
-    size: 9999,
-  })
-
-  const roles = rolesResponse?.content || []
-
-  function handleAddUser(user: UserInput): void {
-    addUser(user, {
-      onSuccess: () => {
-        router.push('/users')
-      },
-    })
-  }
-
-  const onSubmit = handleSubmit(handleAddUser)
-
-  return (
-    <>
-      <Title py="md" order={2}>
-        <Flex>
-          <IconUserPlus size="32" />
-
-          <Text ml="xs">{t('addUpdateUserView.add.title')}</Text>
-        </Flex>
-      </Title>
-
-      <Card shadow="sm" p="lg" radius="sm" withBorder maw="81.25rem">
-        <UserForm
-          title={t('addUpdateUserView.form.userDataHeading')}
-          roles={roles}
-          onSubmit={onSubmit}
-          control={control}
-          formState={formState}
-          setValue={setValue}
-        />
-      </Card>
-    </>
-  )
-}
-
-AddUserView.getLayout = function getLayout(
-  page: React.ReactNode,
-  version?: string,
-): JSX.Element {
-  return (
-    <AuthLayout
-      routeName={getTranslation('addUpdateUserView.add.title')}
-      version={version}
-    >
-      {page}
-    </AuthLayout>
-  )
-}
 
 export const getStaticProps = baseGetStaticProps()
