+++ packages/lib/package.json
@@ -40,6 +40,7 @@
   "devDependencies": {
     "@babel/core": "7.21.0",
     "@emotion/react": "11.10.6",
+    "@emotion/server": "11.11.0",
     "@fortawesome/fontawesome-svg-core": "6.3.0",
     "@fortawesome/free-solid-svg-icons": "6.3.0",
     "@fortawesome/react-fontawesome": "0.2.0",
@@ -51,6 +52,7 @@
     "@mantine/dates": "6.0.2",
     "@mantine/hooks": "6.0.2",
     "@mantine/modals": "6.0.2",
+    "@mantine/next": "6.0.2",
     "@mantine/notifications": "6.0.2",
     "@mantine/spotlight": "6.0.2",
     "@rollup/plugin-commonjs": "24.0.1",
@@ -104,6 +106,7 @@
   },
   "peerDependencies": {
     "@emotion/react": "11.10.6",
+    "@emotion/server": "11.11.0",
     "@fortawesome/fontawesome-svg-core": "6.3.0",
     "@fortawesome/free-solid-svg-icons": "6.3.0",
     "@fortawesome/react-fontawesome": "0.2.0",
@@ -113,6 +116,7 @@
     "@mantine/dates": "6.0.2",
     "@mantine/hooks": "6.0.2",
     "@mantine/modals": "6.0.2",
+    "@mantine/next": "6.0.2",
     "@mantine/notifications": "6.0.2",
     "@mantine/spotlight": "6.0.2",
     "@tabler/icons-react": "2.11.0",
