+++ packages/lib/src/components/index.ts
@@ -22,7 +22,6 @@ export * from './DeleteDialog'
 export * from './Footer'
 export * from './Header'
 export * from './Home'
-export * from './Layout'
 export * from './LoadingSpinner'
 export * from './Login'
 export * from './NavBar'
