+++ packages/lib/tsconfig.base.json
@@ -2,7 +2,7 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "esModuleInterop": true,
-    "module": "ES6",
+    "module": "ES2020",
     "declaration": true,
     "declarationDir": "types",
     "declarationMap": true,
